Fixed sleep crashing the script when running it in some locales.
awk, and by extension the calc function, both return a number with a decimal separator defined by LC_NUMERIC.
In locales such as France and Germany, this meant calc would return a comma separated decimal.
This caused the script to fail when reaching the sleep command, because it can only parse dot separated decimals.
To fix this, awk in the calc function is forced to use LC_NUMERIC=C (portable locale), which does use dot separated decimals.